+++ thermosphere/src/watchpoints.h
@@ -35,7 +35,7 @@ typedef struct WatchpointManager {
 extern WatchpointManager g_watchpointManager;
 
 void initWatchpoints(void);
-DebugControlRegister retrieveSplitWatchpointConfig(u64 addr, size_t size, WatchpointLoadStoreControl direction, bool strict);
-int addWatchpoint(u64 addr, size_t size, WatchpointLoadStoreControl direction);
-int removeWatchpoint(u64 addr, size_t size, WatchpointLoadStoreControl direction);
+DebugControlRegister retrieveSplitWatchpointConfig(uintptr_t addr, size_t size, WatchpointLoadStoreControl direction, bool strict);
+int addWatchpoint(uintptr_t addr, size_t size, WatchpointLoadStoreControl direction);
+int removeWatchpoint(uintptr_t addr, size_t size, WatchpointLoadStoreControl direction);
 int removeAllWatchpoints(void);
